Exclusive Ownership 2 Implemented
1. In the LoadAnswerGraphFromFile function, the _nodes and _edges
containers were updated to use std::unique_ptr instead of raw pointers.
The std::make_unique function was used to create instances of GraphNode
and GraphEdge for insertion into these containers.
2.Changes were made to the iterator comparisons in the
LoadAnswerGraphFromFile function to accommodate the change
to std::unique_ptr.
3.The chatbot was moved using the move function in the ChatLogic
destructor.
4.The edge.get() function was used to obtain the raw pointer from the
unique_ptr and pass it to the AddEdgeToParentNode and
AddEdgeToChildNode methods.
5.The segfault error was fixed by changing the edge unique_ptr to a
shared_ptr and using the shared_from_this() function in the GraphEdge
class.
6.The use of new and delete was avoided for chatbot creation, and
instead, the chatbot was constructed on the stack.